pyproject.toml
[project]
name = "call-agent-vexu"
version = "0.1.0"
description = ""
authors = [
    {name = "BorhanPetgar",email = "borhangpetgar40@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs (==2.4.0)",
    "aiohttp (==3.10.6)",
    "aiohttp-retry (==2.8.3)",
    "aiosignal (==1.3.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.6.0)",
    "async-timeout (==4.0.3)",
    "attrs (==24.2.0)",
    "certifi (==2024.8.30)",
    "charset-normalizer (==3.3.2)",
    "click (==8.1.7)",
    "exceptiongroup (==1.2.2)",
    "fastapi (==0.115.0)",
    "frozenlist (==1.4.1)",
    "h11 (==0.14.0)",
    "idna (==3.10)",
    "multidict (==6.1.0)",
    "pydantic (==2.9.2)",
    "pydantic-core (==2.23.4)",
    "pyjwt (==2.9.0)",
    "python-dotenv (==1.0.1)",
    "requests (==2.32.3)",
    "sniffio (==1.3.1)",
    "starlette (==0.38.6)",
    "twilio (==9.3.2)",
    "typing-extensions (==4.12.2)",
    "urllib3 (==2.2.3)",
    "uvicorn (==0.30.6)",
    "websockets (==13.1)",
    "yarl (==1.12.1)",
    "python-multipart (==0.0.20)",
    "librosa (==0.11.0)",
    "httpx (==0.28.1)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
